chore: an attempt to build everything under dist
since tsup is configured with `clean: true` both entries collide,
when a change is discovered in each it wipes the other entry as they
share the same outDir.

This is an attempt to build both under the same dir so cleanups,
excluding, etc is easy but still have a separation, especially requried
in dev mode.